+++ tests/BlockCypher/Test/Api/TXInputTest.php
@@ -40,11 +40,11 @@ class TXInputTest extends ResourceModelTestCase
             "age": 5197,
             "wallet_name": "alice",
             "wallet_token": "c0afcccdde5081d6429de37d16166ead",
-            "error": "",
-            "errors": []
+            "error": "message",
+            "errors": ["message"]
         }
         */
-        return '{"prev_hash":"c719e0c52f63d9afbb72b00324499c0510672fa63c205db982188161ee3f105c","output_index":0,"script":"4830450221008627dbea1b070e8ceb025ab0ecb154227a65a34e6e8cd64966f181ca151d354f022066264b1930ad9e638f2853db683f5f81059e8c547bf9b4512046d2525c170c0b01210274cb62e999bdf96c9b4ef8a2b44c1ac54d9de879e2ee666fdbbf0e1a03090cdf","output_value":50000,"sequence":4294967295,"addresses":["n3D2YXwvpoPg8FhcWpzJiS3SvKKGD8AXZ4"],"script_type":"pay-to-pubkey-hash","age":5197,"wallet_name":"alice","wallet_token":"c0afcccdde5081d6429de37d16166ead","error":"","errors":[]}';
+        return '{"prev_hash":"c719e0c52f63d9afbb72b00324499c0510672fa63c205db982188161ee3f105c","output_index":0,"script":"4830450221008627dbea1b070e8ceb025ab0ecb154227a65a34e6e8cd64966f181ca151d354f022066264b1930ad9e638f2853db683f5f81059e8c547bf9b4512046d2525c170c0b01210274cb62e999bdf96c9b4ef8a2b44c1ac54d9de879e2ee666fdbbf0e1a03090cdf","output_value":50000,"sequence":4294967295,"addresses":["n3D2YXwvpoPg8FhcWpzJiS3SvKKGD8AXZ4"],"script_type":"pay-to-pubkey-hash","age":5197,"wallet_name":"alice","wallet_token":"c0afcccdde5081d6429de37d16166ead","error":"message","errors":["message"]}';
     }
 
     /**
@@ -70,6 +70,9 @@ class TXInputTest extends ResourceModelTestCase
         $this->assertNotNull($obj->getWalletName());
         $this->assertNotNull($obj->getWalletToken());
 
+        $this->assertNotNull($obj->getError());
+        $this->assertNotNull($obj->getErrors());
+
         $this->assertEquals(self::getJson(), $obj->toJson());
         return $obj;
     }
@@ -93,5 +96,8 @@ class TXInputTest extends ResourceModelTestCase
         // See http://dev.blockcypher.com/#creating-transactions
         $this->assertEquals($obj->getWalletName(), "alice");
         $this->assertEquals($obj->getWalletToken(), "c0afcccdde5081d6429de37d16166ead");
+
+        $this->assertEquals($obj->getError(), "message");
+        $this->assertEquals($obj->getErrors(), array("message"));
     }
 }