+++ tests/BlockCypher/Test/Api/TXSkeletonTest.php
@@ -24,6 +24,8 @@ class TXSkeletonTest extends ResourceModelTestCase
         $this->assertNotNull($obj->getTosign());
         $this->assertNotNull($obj->getSignatures());
         $this->assertNotNull($obj->getPubkeys());
+        $this->assertNotNull($obj->getError());
+        $this->assertNotNull($obj->getErrors());
 
         $this->assertEquals(self::getJson(), $obj->toJson());
 
@@ -99,13 +101,13 @@ class TXSkeletonTest extends ResourceModelTestCase
             "pubkeys":[
                 "0274cb62e999bdf96c9b4ef8a2b44c1ac54d9de879e2ee666fdbbf0e1a03090cdf"
             ],
-            "error": "",
-            "errors": []
+            "error": "message",
+            "errors": ["message"]
         }
         */
 
         $tx = TXTest::getJson();
-        return '{"tx":' . $tx . ',"tosign":["7e6a71a683f303b6a659daa8009c81c47edd2b14f59b938cb31f8ef2a3e129f5"],"signatures":["30450221008627dbea1b070e8ceb025ab0ecb154227a65a34e6e8cd64966f181ca151d354f022066264b1930ad9e638f2853db683f5f81059e8c547bf9b4512046d2525c170c0b"],"pubkeys":["0274cb62e999bdf96c9b4ef8a2b44c1ac54d9de879e2ee666fdbbf0e1a03090cdf"],"error":"","errors":[]}';
+        return '{"tx":' . $tx . ',"tosign":["7e6a71a683f303b6a659daa8009c81c47edd2b14f59b938cb31f8ef2a3e129f5"],"signatures":["30450221008627dbea1b070e8ceb025ab0ecb154227a65a34e6e8cd64966f181ca151d354f022066264b1930ad9e638f2853db683f5f81059e8c547bf9b4512046d2525c170c0b"],"pubkeys":["0274cb62e999bdf96c9b4ef8a2b44c1ac54d9de879e2ee666fdbbf0e1a03090cdf"],"error":"message","errors":["message"]}';
     }
 
     /**
@@ -118,6 +120,8 @@ class TXSkeletonTest extends ResourceModelTestCase
         $this->assertEquals($obj->getTosign(), array("7e6a71a683f303b6a659daa8009c81c47edd2b14f59b938cb31f8ef2a3e129f5"));
         $this->assertEquals($obj->getSignatures(), array("30450221008627dbea1b070e8ceb025ab0ecb154227a65a34e6e8cd64966f181ca151d354f022066264b1930ad9e638f2853db683f5f81059e8c547bf9b4512046d2525c170c0b"));
         $this->assertEquals($obj->getPubkeys(), array("0274cb62e999bdf96c9b4ef8a2b44c1ac54d9de879e2ee666fdbbf0e1a03090cdf"));
+        $this->assertEquals($obj->getError(), "message");
+        $this->assertEquals($obj->getErrors(), array("message"));
     }
 
     /**