+++ tests/BlockCypher/Test/Common/BlockCypherBaseModelTest.php
@@ -3,6 +3,7 @@
 // Do not add namespace
 
 use BlockCypher\Common\BlockCypherBaseModel;
+use BlockCypher\Test\Api\ErrorTest;
 
 /**
  * Test class for BlockCypherBaseModelTest.
@@ -17,6 +18,28 @@ class BlockCypherBaseModelTest extends \PHPUnit_Framework_TestCase
     public static function getObject()
     {
         return new BlockCypherBaseModel(self::getJson());
+    }
+
+    /**
+     * Gets Json String of Object BlockCypherBaseModel
+     * @return string
+     */
+    public static function getJson()
+    {
+        /*
+        {
+          "error":"message",
+          "errors":[
+            {
+              "error":"message"
+            }
+          ]
+        }
+        */
+
+        $error = ErrorTest::getJson();
+
+        return '{"error":"message","errors":[' . $error . ']}';
     }
 
     /**
@@ -37,31 +60,60 @@ class BlockCypherBaseModelTest extends \PHPUnit_Framework_TestCase
     }
 
     /**
-     * Gets Json String of Object BlockCypherBaseModel
-     * @return string
+     * @depends testSerializationDeserialization
+     * @param BlockCypherBaseModel $obj
      */
-    public static function getJson()
+    public function testGetters($obj)
     {
-        /*
-        {
-            "error": "TestError",
-            "errors": [
-                "TestError1",
-                "TestError2"
-            ]
-        }
-        */
-
-        return '{"error":"TestError","errors":["TestError1","TestError2"]}';
+        $this->assertEquals($obj->getError(), "message");
+        $this->assertEquals($obj->getErrors(), array(ErrorTest::getObject()));
     }
 
     /**
      * @depends testSerializationDeserialization
      * @param BlockCypherBaseModel $obj
      */
-    public function testGetters($obj)
+    public function testGetErrorMessages($obj)
     {
-        $this->assertEquals($obj->getError(), "TestError");
-        $this->assertEquals($obj->getErrors(), array("TestError1", "TestError2"));
+        $this->assertEquals(count($obj->getErrors()), count($obj->getErrorMessages()));
+        $this->assertContains("message", $obj->getErrorMessages());
+    }
+
+    /**
+     * @depends testSerializationDeserialization
+     * @param BlockCypherBaseModel $obj
+     */
+    public function testGetAllErrorMessages($obj)
+    {
+        $this->assertEquals(count($obj->getErrors()) + 1, count($obj->getAllErrorMessages()));
+        $this->assertContains("message", $obj->getAllErrorMessages());
+    }
+
+    /**
+     * @depends testSerializationDeserialization
+     * @param BlockCypherBaseModel $obj
+     */
+    public function testAddError($obj)
+    {
+        $newError = new \BlockCypher\Api\Error();
+        $newError->setError("new error");
+
+        $obj->addError($newError);
+
+        $this->assertContains("new error", $obj->getErrorMessages());
+    }
+
+    /**
+     * @depends testSerializationDeserialization
+     * @param BlockCypherBaseModel $obj
+     */
+    public function testRemoveError($obj)
+    {
+        $oldError = new \BlockCypher\Api\Error();
+        $oldError->setError("message");
+
+        $obj->removeError($oldError);
+
+        $this->assertNotContains("message", $obj->getErrorMessages());
     }
 }