+++ release_notes.md
@@ -6,6 +6,7 @@ v1.0.1
 
 * Added TXInput wallet_name and wallet_token properties
 * Added tx creation sample using a wallet
+* Changed class for BlockCypherBaseModel errors attribute
 
 v1.0.0
 ------